Add files via upload
System Task Completion with below points

Using the provided sample JSON, store the data in a Core Data database along with the download status (Pending, Failed, In Progress, and Completed). Display the list by retrieving the data from Core Data.
Provide an option to download the video in the list item. On click of the download button, the video download should start and progress should be displayed in the list item.
Once the download is completed and saved in document directory, update the record in the database as "Completed". 
When clicking on the list item, it should navigate to the details screen where the user has the option to play the video when click on thumbnail along with details (details should be retrieved from the database based on a unique ID).
Without downloading the video, if the user clicks on the list item, show an alert stating that the video is not downloaded.
Once the download is completed, it shouldn’t download again.